+++ ErsatzTV.Core/Scheduling/PlayoutBuildWarnings.cs
@@ -7,9 +7,13 @@ public class PlayoutBuildWarnings
         TailFillerTooLong += warnings.TailFillerTooLong;
         MidRollContentWithoutChapters += warnings.MidRollContentWithoutChapters;
         DurationFillerSkipped += warnings.DurationFillerSkipped;
+
+        BlockItemSkippedEmptyCollection += warnings.BlockItemSkippedEmptyCollection;
     }
 
     public int TailFillerTooLong { get; set; }
     public int MidRollContentWithoutChapters { get; set; }
     public int DurationFillerSkipped { get; set; }
+
+    public int BlockItemSkippedEmptyCollection { get; set; }
 }