+++ CHANGELOG.md
@@ -52,6 +52,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix UI error editing watermarks and graphics elements on blocks
 - Fix showing playout build failure details when resetting a playout
 - Fix scheduling auto-generated trakt list playlists that contain shows
+- Fix playout builder getting stuck (forever) on block item with an empty collection
+- Fix HLS Direct playback when using custom stream selector or preferred audio language/title
+- Fix selecting embedded subtitles (text and picture) with HLS Direct
 
 ### Changed
 - Do not use graphics engine for single, permanent watermark